Move Push non_blicking into correct module
